:sparkles: [#157] Relocate shared test utilities
