+++ tests/test_logout.py
@@ -5,10 +5,10 @@ import pytest
 from requests import Session
 
 from mozilla_django_oidc_db.models import OIDCClient
+from mozilla_django_oidc_db.tests.utils import keycloak_login
 from mozilla_django_oidc_db.utils import do_op_logout
 
 from .conftest import oidcconfig
-from .utils import keycloak_login
 
 
 @pytest.fixture
